% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapezoid.R
\name{trapezoid}
\alias{trapezoid}
\title{Title}
\usage{
trapezoid(f, range, n)
}
\arguments{
\item{f}{a function with a single argument x}

\item{range}{a vector specifying the lower ans upper end points of the range of integration(a and b),and}

\item{n}{the number of intervals for the Riemann approximation}
}
\value{
a single value which is the Riemann approximation to integral the f(x) from b to a.
}
\description{
Title
}
\examples{
 f = function(x) x^4+x^2
trapezoid(f,c(4,6),1000)
}
